refactor: move classic battle state handlers
