+++ src/helpers/classicBattle/stateHandlers/matchDecisionEnter.js
@@ -1,0 +1,10 @@
+/**
+ * onEnter handler for `matchDecision` state.
+ *
+ * @returns {Promise<void>}
+ * @pseudocode
+ * 1. Placeholder for match decision logic.
+ */
+export async function matchDecisionEnter() {}
+
+export default matchDecisionEnter;
