+++ src/helpers/classicBattle/stateHandlers/matchOverEnter.js
@@ -1,0 +1,14 @@
+import { emitBattleEvent } from "../battleEvents.js";
+
+/**
+ * onEnter handler for the `matchOver` state.
+ *
+ * @returns {Promise<void>}
+ * @pseudocode
+ * 1. Emit `matchOver` battle event.
+ */
+export async function matchOverEnter() {
+  emitBattleEvent("matchOver");
+}
+
+export default matchOverEnter;
