+++ src/helpers/classicBattle/stateHandlers/waitingForPlayerActionExit.js
@@ -1,0 +1,14 @@
+import { emitBattleEvent } from "../battleEvents.js";
+
+/**
+ * onExit handler for `waitingForPlayerAction` state.
+ *
+ * @returns {Promise<void>}
+ * @pseudocode
+ * 1. Disable stat buttons.
+ */
+export async function waitingForPlayerActionExit() {
+  emitBattleEvent("statButtons:disable");
+}
+
+export default waitingForPlayerActionExit;
